background.c         : fixed annoying bug that caused abort trap in rhf1d rh15d_mpi/rh15d.c    : added warning that this code is not up to date rh15d_mpi/writeRay.c : uncommented writing of source function components rhf15d/formal.c      : put Nspace as long rhf1d/formal.c       : put Nspace as long rhf1d/iter_1d.c      : updated to last version from Han rhf1d/piecewise.c    : now using linear interpolation for source function rhf1d/rhf1d.c        : removed some crud, updated to last version from Han rhf1d/solve1d.c      : fix a bug that prevent compilation with clang

+++ background.c
@@ -297,9 +297,9 @@ void Background(bool_t write_analyze_output, bool_t equilibria_only)
 
   if (atmos.moving || atmos.Stokes) {
     atmos.backgrrecno = 
-      (long *) malloc(2*spectrum.Nspect*atmos.Nrays * sizeof(int));
+      (long *) malloc(2*spectrum.Nspect*atmos.Nrays * sizeof(long));
   } else
-    atmos.backgrrecno = (long *) malloc(spectrum.Nspect * sizeof(int));
+    atmos.backgrrecno = (long *) malloc(spectrum.Nspect * sizeof(long));
 
   /* --- Open output file for background opacity, emissivity,
          scattering --                                 -------------- */
